requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
billiard==3.6.4.0
boto3==1.24.75
botocore==1.27.75
celery==5.2.7
certifi==2022.9.14
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
Deprecated==1.2.13
Django==3.2.13
django-celery-beat==2.2.1
django-cors-headers==3.11.0
django-redis==5.2.0
django-rest-framework==0.1.0
django-ses==3.0.1
django-storages==1.12.3
django-timezone-field==4.2.3
djangorestframework==3.13.1
drf-spectacular==0.22.0
drf-yasg==1.21.3
gunicorn==20.1.0
idna==3.3
importlib-resources==5.7.1
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jmespath==1.0.0
jsonschema==4.4.0
kombu==5.2.4
MarkupSafe==2.1.1
mysqlclient==2.1.0
packaging==21.3
phonenumbers==8.12.46
Pillow==9.1.0
prompt-toolkit==3.0.31
psycopg2-binary==2.9.3
pyparsing==3.0.9
pyrsistent==0.18.1
python-crontab==2.6.0
python-dateutil==2.8.2
pytz==2022.2.1
PyYAML==6.0
redis==4.2.2
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.12
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
zipp==3.8.1
